requirements.txt

starlette
fastapi==0.89.1
uvicorn==0.20.0
python-dotenv==0.21.1
anyio==4.0.0
certifi==2023.7.22
exceptiongroup==1.1.3
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
python-telegram-bot==20.5
sniffio==1.3.0
python-dateutil==2.8.2
pymongo==4.6.1
convert-numbers
# pandas
# shutil
# pydantic
python-dateutil
# openpyxl
